moved facility_adapter to a per-router interface
