+++ app/routers/status/facility_adapter.py
@@ -1,9 +1,6 @@
-from fastapi import Request
 from abc import ABC, abstractmethod
-from .routers.status import models as status_models
-from .routers.account import models as account_models
+from . import models as status_models
 import datetime
-import psij
 
 
 class FacilityAdapter(ABC):
@@ -86,92 +83,4 @@ class FacilityAdapter(ABC):
         self : "FacilityAdapter",
         id : str
         ) -> status_models.Incident:
-        pass
-
-
-    @abstractmethod
-    def get_capabilities(
-        self : "FacilityAdapter",
-        ) -> list[account_models.Capability]:
-        pass
-
-
-    @abstractmethod
-    def get_current_user(
-        self : "FacilityAdapter",
-        request: Request,
-        api_key: str
-        ) -> str:
-        """
-            Decode the api_key and return the authenticated user's id.
-            This method is not called directly, rather authorized endpoints "depend" on it.
-            (https://fastapi.tiangolo.com/tutorial/dependencies/)
-        """
-        pass
-
-
-    @abstractmethod
-    def get_user(
-        self : "FacilityAdapter",
-        request: Request,
-        user_id: str
-        ) -> account_models.User:
-        pass
-
-
-    @abstractmethod
-    def get_projects(
-        self : "FacilityAdapter",
-        request: Request,
-        user: account_models.User
-        ) -> list[account_models.Project]:
-        pass
-
-
-    @abstractmethod
-    def get_project_allocations(
-        self : "FacilityAdapter",
-        request: Request,
-        project: account_models.Project
-        ) -> list[account_models.ProjectAllocation]:
-        pass
-
-
-    @abstractmethod
-    def get_user_allocations(
-        self : "FacilityAdapter",
-        request: Request,
-        user: account_models.User,
-        project_allocation: account_models.ProjectAllocation,
-        ) -> list[account_models.UserAllocation]:
-        pass
-
-    
-    @abstractmethod
-    def submit_job(
-        self: "FacilityAdapter",
-        resource: status_models.Resource, 
-        user: account_models.User, 
-        job: psij.Job,
-    ) -> psij.Job:
-        pass
-
-    
-    @abstractmethod
-    def get_job(
-        self: "FacilityAdapter",
-        resource: status_models.Resource, 
-        user: account_models.User, 
-        job_id: str,
-    ) -> psij.Job:
-        pass
-
-    
-    @abstractmethod
-    def cancel_job(
-        self: "FacilityAdapter",
-        resource: status_models.Resource, 
-        user: account_models.User, 
-        job: psij.Job,
-    ) -> bool:
         pass
